Reorganize and refactor some base functionality
