+++ packages/create-sitecore-jss/src/InitializerFactory.ts
@@ -1,0 +1,15 @@
+import { Initializer } from './common/Initializer';
+import { NextjsInitializer, NextjsStyleguideInitializer } from './initializers/nextjs';
+
+export class InitializerFactory {
+  create(name: string): Initializer | undefined {
+    switch (name) {
+      case 'nextjs':
+        return new NextjsInitializer();
+      case 'nextjs-styleguide':
+        return new NextjsStyleguideInitializer();
+      default:
+        return undefined;
+    }
+  }
+}
