+++ packages/create-sitecore-jss/src/common/Answer.ts
@@ -1,4 +1,3 @@
-import { ParsedArgs } from 'minimist';
 import { Answers } from 'inquirer';
 
 export interface Answer extends Answers {
@@ -9,12 +8,4 @@ export interface Answer extends Answers {
   fetchWith: string;
   hostName: string;
   prefix?: boolean;
-}
-
-export interface Initializer {
-  init: (args: ParsedArgs) => void; // tbd
-}
-
-export interface PackageJsonProperty {
-  [key: string]: string;
 }
