+++ packages/create-sitecore-jss/src/common/Initializer.ts
@@ -1,0 +1,5 @@
+import { ParsedArgs } from 'minimist';
+
+export interface Initializer {
+  init: (args: ParsedArgs) => void; // tbd
+}
