+++ packages/create-sitecore-jss/src/common/steps/index.ts
@@ -1,0 +1,3 @@
+export { installPackages } from './install';
+export { nextSteps } from './next';
+export { diffFiles, merge, transform, transformFilename } from './transform';
