+++ packages/create-sitecore-jss/src/common/test-data/test-app/Test.tsx
@@ -1,0 +1,7 @@
+const Test = (): JSX.Element => (
+  <div>
+    <h1>Test</h1>
+  </div>
+);
+
+export default Test;
