+++ packages/create-sitecore-jss/src/common/test-data/test-app/test-file.ts
@@ -1,0 +1,7 @@
+'test'
+'.\\test-data\\test'
+'GraphQL'
+'SSG'
+'https://cm.jss.localhost'
+
+  "Prerender is SSG"
