+++ packages/create-sitecore-jss/src/common/test-data/test-source.package.json
