+++ packages/create-sitecore-jss/src/common/test-data/test-template.ts
@@ -1,0 +1,11 @@
+'<%- appName %>'
+'<%- destination %>'
+'<%- fetchWith %>'
+'<%- prerender %>'
+'<%- hostName %>'
+
+<% if (prerender === 'SSG') { %>
+  "Prerender is SSG"
+<% } else if (prerender === 'SSR') { %>
+  "Prerender is SSR"
+<% } %>
