+++ packages/create-sitecore-jss/src/common/test-data/test.package.json
