+++ packages/create-sitecore-jss/src/initializers/nextjs/index.ts
@@ -1,2 +1,2 @@
 export { NextjsInitializer } from './app/index';
-export { StyleguideInitializer } from './styleguide/index';
+export { NextjsStyleguideInitializer } from './styleguide/index';
