+++ packages/create-sitecore-jss/src/initializers/nextjs/models.ts
@@ -1,4 +1,4 @@
-import { Answer } from '../../models';
+import { Answer } from '../../common/Answer';
 export interface NextjsAnswer extends Answer {
   prerender: string;
   hostName: string;
