+++ packages/create-sitecore-jss/src/initializers/nextjs/styleguide/index.ts
@@ -1,22 +1,33 @@
 import chalk from 'chalk';
 import path from 'path';
 import { ParsedArgs } from 'minimist';
-import { Initializer } from '../../../models';
-import { openPackageJson, transformFiles } from '../../../shared';
+import { Initializer } from '../../../common/Initializer';
+import { Answer } from '../../../common/Answer';
+import { isJssApp, openPackageJson } from '../../../common/utils/helpers';
+import { transform } from '../../../common/steps/index';
 
-export class StyleguideInitializer implements Initializer {
+export class NextjsStyleguideInitializer implements Initializer {
   async init(args: ParsedArgs) {
+    // set destination to cwd, or read from args
+    args.destination = args.destination || path.resolve(process.cwd());
+    args.post = true;
+
+    let pkg;
+    // read package.json in target destination, check if app is JSS app
+    if (!args.yes) {
+      pkg = openPackageJson(`${args.destination}\\package.json`);
+      isJssApp('nextjs-styleguide', pkg);
+    }
+
+    // derive variables from package.json
     // read the package.json to get the appName
-    const pkg = openPackageJson();
-
-    // derive variables from package.json and cwd
-    args.appName = pkg.config.appName;
-    args.destination = path.resolve(process.cwd());
-    args.appPrefix = pkg.config.prefix || args.appPrefix || false;
-
+    args.appName = pkg?.config.appName || 'default';
+    args.appPrefix = pkg?.config.prefix || args.appPrefix || false;
     const templatePath = path.resolve(__dirname, '../../../templates/nextjs/styleguide');
-    await transformFiles(templatePath, args);
+    await transform(templatePath, (args as unknown) as Answer);
 
-    console.log(chalk.green(`Successfully added styleguide to ${args.appName}!`));
+    if (!args.silent) {
+      console.log(chalk.green(`Successfully added styleguide to ${args.appName}!`));
+    }
   }
 }
