+++ packages/create-sitecore-jss/tsconfig.json
@@ -8,7 +8,6 @@
     "forceConsistentCasingInFileNames": true,
     "lib": [
       "es2017",
-      "dom"
     ],
     "moduleResolution": "node",
     "noFallthroughCasesInSwitch": true,
@@ -34,6 +33,6 @@
     "src/templates/**/*",
     "src/test",
     "src/**/*.test.ts",
-    "src/test-data/**/*",
+    "src/**/test-data/**/*",
   ]
 }
